+++ tests/ui/unnecessary_sort_by.fixed
